refactor(tasks): rename TaskRegistry to ProviderRegistry; drop façade
Complete the registry split. TaskRegistry becomes ProviderRegistry, now
scoped to provider lifecycle, enabled state, the list cache, and provider
lookups — nothing else. The pass-through tasks/completedTasks methods and
the owned queryService are dropped; TaskQueryService and SelectionStore are
constructed and injected by AppComposition, and each caller depends only on
the smaller type it needs (design doc 0005, D1: no façade retained).

TasksTabView and URLSchemeHandler now take an injected TaskQueryService and
call it directly for fan-out; every other consumer and preview is updated to
ProviderRegistry.

Tests: TaskRegistryTests → ProviderRegistryTests (registration/enable/lookup/
authorization only). The query fan-out, Today filter, single-list scope, and
scope→section assertions consolidate into a dedicated TaskQueryServiceTests;
TaskQueryScopeTests is folded in and removed. Each split type now owns its
test file.

After this slice TaskRegistry no longer exists in the tree.

Final code slice of the #404 registry split (design doc 0005, D1 / Finding M).
