+++ app/Taskmato/Views/App/MainWindowView.swift
@@ -27,7 +27,8 @@ struct MainWindowView: View {
   var settings: AppSettings
   var statsViewModel: StatsViewModel
   var selectionStore: TaskSelectionStore
-  var registry: TaskRegistry
+  var registry: ProviderRegistry
+  var queryService: TaskQueryService
   var sidebarSelection: SelectionStore
   @Bindable var nav: MainNavigation
 
@@ -40,6 +41,7 @@ struct MainWindowView: View {
         TasksTabView(
           selectionStore: selectionStore,
           registry: registry,
+          queryService: queryService,
           sidebarSelection: sidebarSelection,
           nav: nav,
           settings: settings
@@ -94,7 +96,7 @@ struct MainWindowView: View {
 #Preview {
   let engine = SessionEngine()
   let settings = AppSettings()
-  let registry = TaskRegistry()
+  let registry = ProviderRegistry()
   return MainWindowView(
     presenter: TimerPresenter(engine: engine, settings: settings),
     engine: engine,
@@ -102,6 +104,7 @@ struct MainWindowView: View {
     statsViewModel: .preview,
     selectionStore: TaskSelectionStore(),
     registry: registry,
+    queryService: TaskQueryService(registry: registry, sorter: TaskSorter()),
     sidebarSelection: SelectionStore(registry: registry),
     nav: MainNavigation(settings: settings)
   )
