+++ app/Taskmato/Views/MenuBar/MenuBarPopoverView.swift
@@ -18,7 +18,7 @@ struct MenuBarPopoverView: View {
   var engine: SessionEngine
   var statsViewModel: StatsViewModel
   var selectionStore: TaskSelectionStore
-  var registry: TaskRegistry
+  var registry: ProviderRegistry
   var nav: MainNavigation
 
   @Environment(\.openSettings) private var openSettings
@@ -126,7 +126,7 @@ struct MenuBarPopoverView: View {
     engine: engine,
     statsViewModel: .preview,
     selectionStore: TaskSelectionStore(),
-    registry: TaskRegistry(),
+    registry: ProviderRegistry(),
     nav: MainNavigation(settings: settings)
   )
 }
